update with transformers==4.38.2

+++ optimum/habana/transformers/models/gemma/modeling_gemma.py
@@ -365,7 +365,7 @@ class GaudiGemmaForCausalLM(GemmaForCausalLM):
 
         hidden_states = outputs[0]
         logits = self.lm_head(hidden_states)
-
+        logits = logits.float()
         loss = None
         if labels is not None:
             # Shift so that tokens < n predict n
